// web/src/api/models/EntityType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */

/**
 * Entity types for audit logging - matches backend EntityType enum
 */
export enum EntityType {
  INTEGRATION = 'INTEGRATION',
  JIRA_WEBHOOK = 'JIRA_WEBHOOK',
  JIRA_WEBHOOK_EVENT = 'JIRA_WEBHOOK_EVENT',
  SITE_CONFIG = 'SITE_CONFIG',
  CACHE = 'CACHE',
}